rename colours to palette to avoid english/american spelling confusion
