+++ src/app/components/ThemeProvider/palette.ts
