+++ src/app/components/ThemeProvider/themes/archive.ts
@@ -1,8 +1,8 @@
-import { ARCHIVE_BLUE, WHITE } from '../colours';
+import { ARCHIVE_BLUE, WHITE } from '../palette';
 import withThemeProvider from '../withThemeProvider';
 
 const archiveTheme = {
-  colours: {
+  palette: {
     BRAND_BACKGROUND: ARCHIVE_BLUE,
     BRAND_LOGO: WHITE,
     BRAND_FOREGROUND: WHITE,
